+++ subwabbit/CHANGELOG.md
@@ -1,14 +1,21 @@
 # subwabbit
 
-## [1.1.0] - 2019-06-13 - 19:16 - Matej Jakimov <matej.jakimov@gmail.com>
+## [2.0.0] - 2019-06-14 15:41 - Matej Jakimov <matej.jakimov@gmail.com>
+### Changed
+- Changed names of formatter methods:
+  - `get_common_line_part` -> `format_common_features`
+  - `get_item_line_part` -> `format_item_features`
+  - `get_vw_line` -> `get_formatted_example`
+ 
+## [1.1.0] - 2019-06-13 19:16 - Matej Jakimov <matej.jakimov@gmail.com>
 ### Changed
 - Improved error message when malformed prediction is returned
 
-## [1.0.1] - 2019-06-04 - 16:08 - Matej Jakimov <matej.jakimov@gmail.com>
+## [1.0.1] - 2019-06-04 16:08 - Matej Jakimov <matej.jakimov@gmail.com>
 ### Fixed
 - documentation improvements
 
-## [1.0.0] - 2019-04-14 - 17:16 - Matej Jakimov <matej.jakimov@gmail.com>
+## [1.0.0] - 2019-04-14 17:16 - Matej Jakimov <matej.jakimov@gmail.com>
 ### Changed
 - open sourced